Refactor: changed location of include files and #include accordingly
